CheXNet-master/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
Pillow>=8.0.1
scikit-learn>=0.23.2
streamlit>=1.22.0
plotly>=5.14.0
